refactor(doctor): split legacy model migrations
